--- frontend/src/components/debiai/statistics/dataAnalysis/dataExport/ExportMenu.vue
@@ -1,225 +1,0 @@
-<template>
-  <div id="ExportMenu">
-    <h2>
-      Export the selected samples
-      <button @click="$emit('cancel')" class="red">Cancel</button>
-    </h2>
-    <!-- Export params -->
-    <form v-on:submit.prevent class="dataGroup">
-      <!-- Selected samples -->
-      <div class="data">
-        <span class="name"> Exported samples </span>
-        <span class="value">
-          {{ selectedData.length }} / {{ data.nbLines }}
-        </span>
-      </div>
-      <!-- Selection Name -->
-      <div class="data">
-        <span class="name">
-          Selection name
-
-          <DocumentationBlock>
-            Name that will be written in the expoted 'selection_name' Json
-            field.
-          </DocumentationBlock>
-        </span>
-        <span class="value">
-          <input type="text" v-model="selectionName" style="flex: 2" required />
-        </span>
-      </div>
-    </form>
-
-    <!-- Export method list -->
-    <div>
-      <h3 style="text-align: left; padding-top: 15px">
-        Select an export method:
-      </h3>
-      <div id="exportMethods" class="card" v-if="exportMethods">
-
-        <div v-if="exportMethods.length === 0" class="marged">
-          No export method available, please create one.
-        </div>
-
-        <div class="itemList">
-          <div class="exportMethod item" v-for="exportMethod in exportMethods" :key="exportMethod.id">
-            <h3 class="name">
-              {{ exportMethod.name }}
-              <button class="red" @click="deleteExportMethod(exportMethod.id)" v-if="exportMethod.deletable !== false">
-                <b>Delete</b>
-              </button>
-              <span style="flex: 1"></span>
-              <button class="green" @click="exportSamples(exportMethod.id)" :disabled="exporting">
-                Export
-              </button>
-            </h3>
-
-            <div class="details">
-              <div class="type">{{ exportMethod.type }}</div>
-              <div class="parameters">
-                <div class="parameter" v-for="(parameter, i) in exportMethod.parameters" :key="i">
-                  <u> {{ exportMethod.parameterNames[i] }}:</u>
-                  {{ parameter }}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div v-else id="exportMethods" class="card" style="justify-content: center; color: grey">
-        Loading export methods
-      </div>
-    </div>
-    <div style="text-align: right">
-      <button class="green" @click="newExportMethod">New export method</button>
-    </div>
-
-    <!-- New Method modal -->
-    <modal v-if="newExportMethodModal">
-      <ExportMethodCreator @cancel="newExportMethodModal = false"
-        @created="newExportMethodModal = false; loadExportMethods()" />
-    </modal>
-  </div>
-</template>
-
-<script>
-import ExportMethodCreator from "./ExportMethodCreator.vue";
-
-export default {
-  components: { ExportMethodCreator },
-  name: "ExportMenu",
-  data() {
-    return {
-      exportMethods: null,
-      selectionName: "DebiAI Selection",
-      exporting: false,
-
-      newExportMethodModal: false,
-    };
-  },
-  props: {
-    data: { type: Object, required: true },
-    selectedData: { type: Array, required: true },
-  },
-  created() {
-    // Load the exportMethods
-    this.loadExportMethods();
-  },
-  methods: {
-    loadExportMethods() {
-      this.exportMethods = null;
-      this.$backendDialog.getExportMethods().then((exportMethods) => {
-        this.exportMethods = exportMethods;
-      });
-    },
-    exportSamples(methodId) {
-      let projectId = this.$store.state.ProjectPage.projectId;
-      let selectedHash = this.selectedData.map(
-        (selectedIndex) => this.data.sampleHashList[selectedIndex]
-      );
-
-      this.exporting = true;
-
-      this.$backendDialog
-        .exportSelection(projectId, this.selectionName, methodId, selectedHash)
-        .then(() => {
-          this.$store.commit("sendMessage", {
-            title: "success",
-            msg: "Selection exported successfully",
-          });
-          this.$emit("exported");
-        })
-        .catch((e) => {
-          console.log(e);
-          if (e.response && e.response.data) {
-            this.$store.commit("sendMessage", {
-              title: "error",
-              msg: e.response.data,
-            });
-          } else {
-            this.$store.commit("sendMessage", {
-              title: "error",
-              msg: "Error while exporting the selection",
-            });
-          }
-        }).finally(() => {
-          this.exporting = false;
-        });
-    },
-
-    deleteExportMethod(id) {
-      this.$backendDialog.deleteExportMethod(id).then(() => {
-        this.$store.commit("sendMessage", {
-          title: "success",
-          msg: "Export method deleted",
-        });
-      }).catch((err) => {
-        console.error(err);
-        this.$store.commit("sendMessage", {
-          title: "error",
-          msg: "Error while deleting the export method",
-        });
-      }).finally(() => this.loadExportMethods());
-
-    },
-
-    newExportMethod() {
-      this.newExportMethodModal = !this.newExportMethodModal;
-    },
-  },
-  computed: {},
-};
-</script>
-
-<style scoped>
-/* Form: */
-.dataGroup {
-  flex-direction: column;
-}
-
-.dataGroup .data+.data {
-  padding-top: 4px;
-}
-
-.dataGroup .value {
-  flex: 1;
-}
-
-/* Methods: */
-#exportMethods {
-  display: flex;
-  min-width: 600px;
-  height: 400px;
-  overflow: auto;
-}
-
-.exportMethod {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: stretch;
-}
-
-.exportMethod .name {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.exportMethod .details {
-  display: flex;
-  align-items: flex-start;
-  padding: 10px;
-  gap: 10px;
-}
-
-.exportMethod .details .parameters {
-  border-left: 1px solid rgb(75, 75, 75);
-  text-align: left;
-  padding-left: 15px;
-  opacity: 80%;
-}
-
-.exportMethod .details .parameter {
-  padding-bottom: 8px;
-}
-</style>
